% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basepreplot.R
\name{basepreplot}
\alias{basepreplot}
\title{Generate the base frequency plot of miRNA}
\usage{
basepreplot(file_dataframe, width = 0.6, size = 12)
}
\arguments{
\item{file_dataframe}{A dataframe. The output result after running mirnapredata.}

\item{width}{A numeric. The width of the output bar plot, and default is 0.6.}

\item{size}{A numeric. The size of axis text, and default is 0.6.}
}
\value{
The miRNA base frequency plot
}
\description{
Generate the base frequency plot of miRNA
}
\examples{
##P1
P1_miRNA_data <- mirnapredata(mirnaseq_dataframe = P1_miRNA_count)
##P2
P2_miRNA_data <- mirnapredata(mirnaseq_dataframe = P2_miRNA_count)
##F1
F1_miRNA_data <- mirnapredata(mirnaseq_dataframe = F1_miRNA_count)
##Drawing
basepreplot(file_dataframe = P1_miRNA_data)
basepreplot(file_dataframe = P2_miRNA_data)
basepreplot(file_dataframe = F1_miRNA_data)
}
