% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{F1_miRNA_rpm}
\alias{F1_miRNA_rpm}
\title{Sequences of miRNAs from one species}
\format{
A dataframe containing all miRNAs with RPM value in "F1" (the S3 generations of Parents (Maternal parent: Triticum turgidum; Male parent: Aegilops tauschii)).
}
\source{
Generated from the S3 generations of Triticum turgidum (AABB) and Aegilops tauschii (DD).
}
\description{
The first column of this data set is all miRNA sequences from "F1", and other columns are the corresponding RPM values of different samples
}
\examples{
data(F1_miRNA_rpm)
}
\references{
Li, A., et al., mRNA and Small RNA Transcriptomes Reveal Insights into Dynamic Homoeolog Regulation of Allopolyploid Heterosis in Nascent Hexaploid Wheat. Plant Cell, 2014. 26(5): p. 1878-1900.
}
