% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic.R
\name{genetic}
\alias{genetic}
\title{Classification for 12 expression patterns}
\usage{
genetic(pv11, pv12, pv21, fc11, fc12, fc21, Pvalue)
}
\arguments{
\item{pv11}{A numeric. The P value of F1_vs_P1 (Treatment:F1; Control:P1).}

\item{pv12}{A numeric. The P value of F1_vs_P2 (Treatment:F1; Control:P2)}

\item{pv21}{A numeric. The P value of P2_vs_P1 (Treatment:P2; Control:P1)}

\item{fc11}{A numeric. The Log2(FoldChange) value of F1_vs_P1 (Treatment:F1; Control:P1)}

\item{fc12}{A numeric. The Log2(FoldChange) value of F1_vs_P2 (Treatment:F1; Control:P2)}

\item{fc21}{A numeric. The Log2(FoldChange) value of P2_vs_P1 (Treatment:P2; Control:P1)}

\item{Pvalue}{A numeric. Filtration criteria of P value for Classification.}
}
\value{
A dataframe.
}
\description{
The input data is generated from the analysis result of DESeq2.
}
