% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirnapredata.R
\name{mirnapredata}
\alias{mirnapredata}
\title{Generate the data of miRNA base frequence in each position}
\usage{
mirnapredata(mirnaseq_dataframe)
}
\arguments{
\item{mirnaseq_dataframe}{A dataframe. The first column must be the sRNA sequence.}
}
\value{
A dataframe. About the output results, the first column is the base, the second column is the base frequency, the third column is the position.
}
\description{
Generally, the "T" base account for the highest percentage of miRNA in the first position.The function of "mirnapredata" can provide the input data for the next drawing of miRNA base distribution in each position.
}
\examples{
##P1
P1_miRNA_data <- mirnapredata(mirnaseq_dataframe = P1_miRNA_count)
##P2
P2_miRNA_data <- mirnapredata(mirnaseq_dataframe = P2_miRNA_count)
##F1
F1_miRNA_data <- mirnapredata(mirnaseq_dataframe = F1_miRNA_count)
}
