% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{P1_miRNA_count}
\alias{P1_miRNA_count}
\title{Sequences of miRNAs from one species}
\format{
A dataframe containing all miRNAs with count value in "P1" (Male parent: Aegilops tauschii).
}
\source{
Generated from the Aegilops tauschii (DD).
}
\description{
The first column of this data set is all miRNA sequences from "P1", and other columns are the corresponding count values from different samples
}
\examples{
data(P1_miRNA_count)
}
\references{
Li, A., et al., mRNA and Small RNA Transcriptomes Reveal Insights into Dynamic Homoeolog Regulation of Allopolyploid Heterosis in Nascent Hexaploid Wheat. Plant Cell, 2014. 26(5): p. 1878-1900.
}
