% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{P1_sRNA_seq}
\alias{P1_sRNA_seq}
\title{Sequences of sRNAs from one species}
\format{
A dataframe containing numerous sequences of all sRNAs in "P1" (Male parent: Aegilops tauschii). However, only 400 sRNAs are selected as test data due to the large data of sRNA.
}
\source{
Generated from the Aegilops tauschii (DD).
}
\description{
The first column of this data set is all sequences from one speceie for the data statistics
}
\details{
This \code{data.frame} is very useful for the functional demonstration of "srnapredata"
}
\examples{
data(P1_sRNA_seq)
}
\references{
Li, A., et al., mRNA and Small RNA Transcriptomes Reveal Insights into Dynamic Homoeolog Regulation of Allopolyploid Heterosis in Nascent Hexaploid Wheat. Plant Cell, 2014. 26(5): p. 1878-1900.
}
