% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{P2_sRNA_seq}
\alias{P2_sRNA_seq}
\title{Sequences of sRNAs from one species}
\format{
A dataframe containing numerous sequences of all sRNAs in "P2" (Maternal parent: Triticum turgidum).
}
\source{
Generated from the Triticum turgidum (AABB).
}
\description{
The first column of this data set is all sequences from one species for the data statistics
}
\details{
This \code{data.frame} is very useful for the functional demonstration of "srnapredata"
}
\examples{
data(P2_sRNA_seq)
}
\references{
Li, A., et al., mRNA and Small RNA Transcriptomes Reveal Insights into Dynamic Homoeolog Regulation of Allopolyploid Heterosis in Nascent Hexaploid Wheat. Plant Cell, 2014. 26(5): p. 1878-1900.
}
